[Re-Architect for Multiple Citation Data Sources] Created Mock Data Source; add Data Source Interface methods; added configuration for whether to use test/live sources
