+++ src/main/java/svc/data/citations/CitationDataSource.java
@@ -1,5 +1,12 @@
 package svc.data.citations;
 
-public interface CitationDataSource {
+import svc.models.Citation;
 
+import java.util.Date;
+import java.util.List;
+
+public interface CitationDataSource {
+    Citation getByCitationNumberAndDOB(String citationNumber, Date dob);
+    List<Citation> getByLicenseAndDOB(String driversLicenseNumber, Date dob);
+    List<Citation> getByNameAndMunicipalitiesAndDOB(String lastName, List<Long> municipalities, Date dob);
 }
