+++ src/main/java/svc/data/citations/CitationDataSourceFactory.java
@@ -1,8 +1,11 @@
 package svc.data.citations;
 
 import com.google.common.collect.Lists;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Component;
 import svc.data.citations.datasources.CITATION_DATASOURCE;
+import svc.data.citations.datasources.mock.MockCitationDataSource;
 import svc.data.citations.datasources.tyler.TylerCitationDataSource;
 import svc.data.jdbc.BaseJdbcDao;
 import svc.logging.LogSystem;
@@ -12,15 +15,25 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.StringJoiner;
 
+@Component
 public final class CitationDataSourceFactory extends BaseJdbcDao {
+    @Value("${stlcourts.citationDataSources.testEnabled}")
+    Boolean testCitationSourcesEnabled;
+
+    @Value("${stlcourts.citationDataSources.liveEnabled}")
+    Boolean liveCitationSourcesEnabled;
+
+    private static final CitationDataSource mockSource = new MockCitationDataSource();
     private static final CitationDataSource tylerSource = new TylerCitationDataSource();
 
-    public static List<CitationDataSource> getAllCitationDataSources() {
-        return Lists.newArrayList(tylerSource);
+    public List<CitationDataSource> getAllCitationDataSources() {
+        return Lists.newArrayList(getEnabledSources());
     }
 
     public List<CitationDataSource> getCitationDataSourcesForMunicipalities(List<Long> municipalityIds) {
         List<CITATION_DATASOURCE> sourcesForMunicipalities = getCitationDataSourceNames(municipalityIds);
+
+        //TODO: MOck
 
         List<CitationDataSource> dataSources = new ArrayList<>();
 
@@ -34,6 +47,20 @@ public final class CitationDataSourceFactory extends BaseJdbcDao {
         }
 
         return dataSources;
+    }
+
+    private List<CitationDataSource> getEnabledSources() {
+        List<CitationDataSource> sources = Lists.newArrayList();
+
+        if(testCitationSourcesEnabled) {
+            sources.add(mockSource);
+        }
+
+        if(liveCitationSourcesEnabled) {
+            sources.add(tylerSource);
+        }
+
+        return sources;
     }
 
     private List<CITATION_DATASOURCE> getCitationDataSourceNames(List<Long> municipalityIds) {
